Fix -Wshorten-64-to-32: config.hpp and config.cpp
Use std::size_t instead of unsigned int for container sizes.

There's one place where it still uses signed int, for using a negative index to
access children from the end of the array. Overflow shouldn't be a worry here,
as it would require 2**31 objects to be in the container, implying the container
and children would use at least 32GB even if every child object was empty.

Co-authored by: Pentarctagon