Fix -Wshorten-64-to-32: math.hpp
The last uses of this were removed in 0fbc12ea01986bc97ef6426e0edc89f12224a0d5.

The only uses of the other fixed-point functions from math.hpp are in
src/sdl/utils.cpp. That should probably move to using color.hpp instead,
but this commit is just to solve the -Wshorten-64-to-32 warning.