filter true damage and attacks in [filter_attack/weapon] except in [damage] and [attacks]
Now that [event][filter_attack] can manage the activity of special weapons, the question arises of the possibility or not of filtering the damage and the number of attacks taking into account the modifications made by [damage] and [attacks].

In this PR, the real damages and attacks are taken into account except for the special ones in order to avoid problems of infinite loops.

I admit, it was also while looking for this solution for damage_type that I came to propose it for two filterable parameters in special weapons, which had led to filtering only the unchanged parameters.