gui: Handle SDL_WINDOWEVENT_SIZE_CHANGED
Instead of SDL_WINDOWEVENT_RESIZED.

The documentation is a bit unclear, but SDL_WINDOWEVENT_RESIZED seems
not to trigger under certain circumstances, specifically when changing
resolution via SDL_SetWindowSize(); SDL_WINDOWEVENT_SIZE_CHANGED
meanwhile is implied to trigger any time the window size changes
whether it's from an API call, or the operating system, or a user
action.

<https://wiki.libsdl.org/SDL2/SDL_WindowEventID>

We need to translate at least one of these events into a GUI2 event in
order for dialogs to be properly re-laid out after the window size
changes, but we should not handle both because that could result in
unnecessary work recalculating all dialog layouts.

Closes #7436.