gui2/title_screen: Fix Version label translation not updating
This unifies both the Version and the Language menu labels into a
single method for both that is called both during construction and
after switching languages. Both have this issue where they aren't
t_string instances, so changing languages has no effect on them for
that reason and they need to be manually re-translated.