Implement one part of the -Wshorten-64-to-32 changes from #7451
Mainly since it's particularly non-obvious why it currently works for `generic_decode_be()` in the case when `in.find_last_not_of("=")` doesn't find anything:
* It returns npos, which is the max value of an std::size_t
* this turns into -1 when assigned to an int
* -1 + 1 equals 0
* 0 * 8 / 6 is still 0
* `reserve(0)` is pointless, but at least valid
This is not a behavior change because it already returns nothing after it does all its work if `out.size()` doesn't equal 0 in this case.

And for `generic_decode_le()` in the case when `in.find_last_not_of("=")` doesn't find anything:
* It returns npos, which is the max value of an std::size_t
* this turns into -1 when assigned to an int
* -1 * 8 / 6 is -0.75, which truncates to 0 because this is integer division
* `reserve(0)` is pointless, but at least valid
This is not a behavior change since the for-loop starts i at 0 and then checks that `i <= last_char`, however last_char is -1 so the loop is just skipped.