UtBS: Make the name of the dark assassin translatable again
It needs to be translatable because it needs to match the other strings that
call him by name. It was translatable until 1.15.4, and several of the .po
files that are already in the 1.16 still include it as an obsolete message, so
the next pot-update run will turn this back into a translated text in
Bulgarian, Greek, Irish, Galician, Latin, Lithuanian, Russian, Serbian and
Traditional Chinese.