Refactor hotkey_handler::do_execute_command
previously, when showing a menu, the hotkey
handler would generate menu items, and then
track the index they are in, which means they
had to keep track of which items were inserted
at which position, which made the code rather
complicated.

Now the hoteky handler generates a informative
command id string for each of those those
generated menu items which is then passed back
to do_execute_command.

To do this we change, the parameter of
do_execute_command and related functions to use
a new struct type, (the parameter used to be
just a HOTKEY_COMMAND enum, then (when wml
command hoeky were added) it changes to
hotkey_command structs, but didn't really match
the uses of that function since is also used with
menu items that aren't hotkey commands)

The new struct is also easier to pass around and
should also make future refactors of this
funcction easier.

With this there are in particular no longer two different
codepaths for wml menu items fired via hotkeys and wml menu
items fired via the right click menu.

This removed also a few codes that i think are now unnecessary,
in particular for old markup, and codes that uses the "id" for label
(the code that generates the menu items now simply sets the "label"
attribute).