Remove an explicit GETTEXT_DOMAIN "wesnoth"
C++ source files are considered to be in textdomain "wesnoth" by default.

The 1.16 CMake scripts that sort files into textdomains miss this file,
because it treats files with no `GETTEXT_DOMAIN` as being in "wesnoth",
but not ones that explicitly declare themselves in "wesnoth".

That's not an issue with the CMake scripts in master, I'm just cherry
picking the commit so that there isn't a "missing" commit on master.
It wasn't an issue with the 1.16 SCons scripts either.

(cherry picked from commit f0ceaf2ae086bf412e56778656e3f4b95824a261)