Remove an explicit GETTEXT_DOMAIN "wesnoth"
C++ source files are considered to be in textdomain "wesnoth" by default.

The 1.16 CMake scripts that sort files into textdomains miss this file,
because it treats files with no `GETTEXT_DOMAIN` as being in "wesnoth",
but not ones that explicitly declare themselves in "wesnoth".

Doing a pot-update with SCons includes this file anyway, so the official
"pot-update and regenerate doc files" commits already include it, even
without this fix.