UtBS: improve po hints
po: The rest of this commit message is for the translators' changelog, because it's notes about
what changed recently, as found by looking at fuzzies while updating the German translation.

S01 "Come on, Kaleh, we have to go see if anyone is hurt or needs help. ..." was just a grammar fix,
of "answer to the call".

UtBS S08's dialogue had Esanoo refer to Melusand as "my master" repeatedly. That changed to
"leader" or "wise leader" in 1.15.4, mainly without other changes in the text.

Other lines should have changed "master" to "leader". Will be fixed in 1.17 but not backported:
* S08 Zhul's "You don’t know where to find your master?"
* S08 Kaleh's two game-over lines because "too many merfolk have died"

UtBS S08 "You dare defy me?!..." : the change is just a typo fix

UtBS S10 "This will go much faster if you don’t interrupt me. ..." the one-word change is from "the
many" to "many" in the last sentence.

UtBS unit type Dawarf: the unit description had a grammar correction of "it's isn't"