WC: change some messages when units join, and add po hints
Add the po hints that have been added to 1.16.

Drake Fighters now say "blade" instead of "sword".

Remove the response when a human finds an Ulf, because it didn't sound right as
the first thing a human says to a dwarf. The Ulf’s "a good day to die" could be
read as a Native American battlecry, and "brave words" as "words of an Indian
Brave", leaving the "little big man" to sound like a name.

Without catching the culture reference, it came across as a jibe about the
dwarf's height.

Catching the culture reference, the logic seems reversed. Why would another
dwarf use the generic form, but a human feel comfortable saying that?