Offscreen rendering for windows.
Windows will now render to an offscreen buffer. They only render when
something changes, and the re-rendered area is kept to a minimum.

When the window needs to be redrawn to the screen, it simply copies
the appropriate region from the offscreen buffer.

This should greatly improve performance in cases where a window isn't
changing its contents, but needs to be redrawn frequently. For example
a pop-up dialog being drawn over animated terrain (see #7615).