Improve performance of the game history query by a lot.
Fixes #7988

I have no idea why this helps anything at all, since my understanding has always been that inner joins vs a big where clause are essentially two different ways of accomplishing the same thing, but changing the two inner joins to a where clause improves query performance from 6.3 to 0.012 seconds (when running the query from the website VM). As best I can tell from the explain plans, the inner joins are somehow confusing the query optimizer to use the game_content_info table as the starting point of the query instead of the game_info table, but why it's doing that or why it has such a large impact remains a mystery.