UtBS: Units with tailwind have green orbs and rest heal
In UtBS, at the start of the player's turn any units that are next to a
Quenoth Shaman get +1 move (or +2 for some of the L2/L3 shamans).

Without this fix, the buffed units have an orange orb instead of the green
unmoved orb, and don't rest heal next turn (unless the player moves them a
little to make the moves remaining exactly match the max).

The buffed units can move a little and still rest heal; this makes the buff
a little stronger, and the lore explanation is that the shamans can fix the
wounds well enough that a little movement is okay.