use the tag_name of each special to avoid infinite loops.
With this method, less cumbersome than the previous proposal that I removed, the check works perfectly for special weapons and for the event filter while the proposal to block at the level of the second loop creates a gap between the two cases, we could had a special train active but the filter_attack did not match or vice versa.

I arranged so that the tag_name is only used for example for [filter_self] when applied to self, so the type of a modified opponent can be detected by a [filter_opponent] of a [damage] applied to self.