Help Browser: only show one image if male and female are the same
If a unit type has male and female versions, then two images of the unit are
displayed at the top-left of the help page. However, it did this even if the
two images were the same, which made the duplication look like a bug.

The Naga Fighter is one of the affected unit types.