New unit test for UtBS-style [heals] with a filter
The ability's code in UtBS hasn't changed since 1.16, but there's now a bug about
healing sometimes not working in UtBS S02. The new test was written to check
for an engine bug when a healer has two units next to it, only one of which
passes the filter; although it's passing, it seems reasonable to keep the test.

Also, use `{ASSERT ...}` for better logging in the old healing test.