reverse change to filter_ability ->experimental_filter_ability
By changing the names of filter_ability to experimental_filter_ability, I think I'm making a mistake. By doing this, we will force the authors who use these filters to change the names when we exit the "experimental" phase and I don't like that, it would be better to just indicate it in the changelog and the wiki.