UtBS: Mention the campaign mechanics in the campaign description
Let players see what to expect when looking through the campaign menu, instead
of only giving players this information after they start the campaign and have
no clear or quick way to return to the main or campaign selection menus.

The preceding translatable string includes trailing newlines, but is the same
as 1.16. So I've left that as-is, while using a separate non-translatable
string for the new string's newlines.