08_Clearwater_Port.map
Fixed starting location (bug)

In 1.17, the author has changed the enemy side numbers in the scenario file. However, it seems that the map file starting locations where not updated accordingly. This file updates the starting location.
However, there is a chance the author did alter the map as whole, and in that case, we should wait for his upload for the new file.