wesnothd: update current turn in init_side instead of end_turn
This matches the behavior on the client where the events that
happen after end_turn (the "turn end" wml events) happen during
the last sides turn not the next sides turn. In particular this
fixes #2563. It shodul also make it easier to fix #3899 later.
It also changes how descption is updated, since carrying that
return value arround specifying whether it was updated was becoming
annoying.