Use argparse for add_source_file
Treat an argument of "--help" as a request to print help, instead of a filename
to add to the source lists.

Treat the arguments "src/foo.cpp" "src/foo.hpp" as two filenames, instead of
adding src/foo.cpp to target src/foo.hpp. The old command line

    ./add_source_file a b c

becomes

    ./add_source_file a --target b --target c

When saying that pbxproj is required, suggest optionally using a venv instead
of installing systemwide.